refactor: No panic, I/O to util
